Moves and fixes RPhenoscape vignette so it renders
This moves the RPhenoscape vignette to using `01_` as a prefix, resulting
in it getting rendered first. Otherwise, `pkgdown::build_site()` doesn't
get to it because the PARAMO vignette fails.

Also fixes it up by including the `library(rphenoscape)` command, as
otherwise all the package's function would need to be prefixed.